Extract GTFSCSVConverterHook for testing
* Remove byte order mark from name
